pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "killswitch-engage"
version = "1.0.0"
description = "Sistema inteligente de recomendação de jogos Steam"
authors = [{ name = "Isaque" }]
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "slowapi>=0.1.9",
    "python-multipart==0.0.6",
    "httpx==0.25.0",
    # Database / Cache
    "asyncpg==0.29.0",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    # Config
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    # ML
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "joblib>=1.3.0",
    "lightfm>=1.17",
    "hdbscan>=0.8.33",
    "torch>=2.0.0",
    # MLOps
    "mlflow>=2.5.0",
    "optuna>=3.3.0",
    # Viz / Stats
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pingouin>=0.5.3",
    # Utils
    "tqdm>=4.66.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",      # async test client
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "bandit>=1.7.5",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 99
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 99
known_first_party = ["src"]
known_third_party = ["fastapi", "pydantic", "sqlalchemy", "asyncpg", "redis", "sklearn", "lightfm", "torch", "mlflow", "optuna"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = false
warn_unused_ignores = true
no_implicit_optional = true

[tool.flake8]
max-line-length = 99
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", "data/", "models/", "mlruns/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "requires_db: marks tests that need a real database connection (deselect with -m 'not requires_db')",
    "slow: marks slow integration tests",
]

[tool.bandit]
exclude_dirs = ["tests", "scripts", "notebooks"]
skips = ["B101"]   # assert statements OK in tests

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
